+++ src/app/modules/_unused/experimental/student-monitoring/student-detail-option/student-detail-option.component.scss
